pyproject.toml
[tool.poetry]
name = "langgraph-simple-chatbot"
version = "0.1.0"
description = ""
authors = ["Allwyn Siluvainathan <asiluvai@akamai.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^24.10.0"
langchain-openai = "^0.2.10"
langchain = "^0.3.9"
langgraph = "^0.2.53"
langchain-community = "^0.3.8"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
